requirements.txt
streamlit==1.56.0
google-generativeai==0.8.6
python-dotenv==1.2.2
